fixed vram leak, and added better framebuffer management

+++ src/me/anno/fonts/mesh/FontMeshBase.kt
@@ -5,9 +5,9 @@ import me.anno.objects.cache.CacheData
 import org.joml.Matrix4fArrayList
 
 abstract class FontMeshBase: CacheData {
-    var minX = Double.POSITIVE_INFINITY
-    var minY = Double.POSITIVE_INFINITY
-    var maxX = Double.NEGATIVE_INFINITY
-    var maxY = Double.NEGATIVE_INFINITY
+    var minX = Float.POSITIVE_INFINITY
+    var minY = Float.POSITIVE_INFINITY
+    var maxX = Float.NEGATIVE_INFINITY
+    var maxY = Float.NEGATIVE_INFINITY
     abstract fun draw(matrix: Matrix4fArrayList, drawBuffer: (StaticFloatBuffer) -> Unit)
 }
